pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bach"
version = "0.1.0"
description = "Manages tasks in Google Ads"
authors = [
  {name = "Google Inc. (gTech gPS CSE team)", email = "no-reply@google.com"},
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "google-ads-api-report-fetcher[pandas]",
  "croniter",
  "sqlalchemy>2",
  "garf_core",
  "garf_youtube_data_api",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]
[project.optional-dependencies]
server = [
  "fastapi[standard]"
]
all = [
  "bach[server]"
]
[options.extras_require]
test =  [
  "pytest",
  "pytest-cov"
]
[tool.pytest.ini_options]
testpaths = [
  "tests",
]

[project.scripts]
bach="bach.entrypoints.cli:main"
